refactor: move `RegistryPackage` to `index` module
`RegistryPackage` is a line in a raw registyr index file. This should
belong to `index.rs` module.
